Organize Chrome Custom Tab components
