+++ CorePayments/src/main/java/com/paypal/android/corepayments/chromecustomtabs/CustomTabsAdapter.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.corepayments.browserswitch
+package com.paypal.android.corepayments.chromecustomtabs
 
 import android.content.Context
 import android.net.Uri
@@ -9,6 +9,7 @@ import androidx.browser.customtabs.CustomTabsServiceConnection
 import androidx.browser.customtabs.CustomTabsSession
 import androidx.browser.customtabs.EngagementSignalsCallback
 
+/** Defines the Custom Tabs operations required to launch tabs and track browser sessions. */
 internal interface CustomTabsAdapter {
     fun getPackageName(context: Context): String?
     fun bind(context: Context, packageName: String, connection: CustomTabsServiceConnection): Boolean
