+++ CorePayments/src/main/java/com/paypal/android/corepayments/chromecustomtabs/PreparedCustomTab.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.corepayments.browserswitch
+package com.paypal.android.corepayments.chromecustomtabs
 
 import androidx.browser.customtabs.CustomTabsSession
