+++ CorePayments/src/main/java/com/paypal/android/corepayments/chromecustomtabs/TrackedChromeCustomTabResult.kt
@@ -1,6 +1,7 @@
-package com.paypal.android.corepayments.browserswitch
+package com.paypal.android.corepayments.chromecustomtabs
 
 import androidx.annotation.RestrictTo
+import com.paypal.android.corepayments.browserswitch.BrowserSwitchSession
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class TrackedChromeCustomTabResult(
